refactor: Resolving some change request
