+++ src/pages/home/components/ui/Calendar.tsx
@@ -1,0 +1,54 @@
+import { FormControl } from "@/components/ui/form";
+import { Button } from "@/components/ui/button";
+import { CalendarIcon } from "@radix-ui/react-icons";
+import { Calendar } from "@/components/ui/calendar";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
+import { format } from "date-fns";
+import { cn } from "@/lib/utils";
+
+interface FieldProps {
+  value: Date | undefined; // Adjust this to match the Date type or undefined
+  onChange: (value: Date | undefined) => void; // Update the onChange function to accept Date or undefined
+}
+
+const PopoverComponent: React.FC<{ field: FieldProps }> = ({ field }) => {
+  return (
+    <Popover>
+      <PopoverTrigger asChild>
+        <FormControl>
+          <Button
+            variant={"outline"}
+            className={cn(
+              "w-full pl-3 text-left font-normal",
+              !field.value && "text-muted-foreground",
+            )}
+          >
+            {field.value ? (
+              format(field.value, "PPP")
+            ) : (
+              <span>Pick a date</span>
+            )}
+            <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+          </Button>
+        </FormControl>
+      </PopoverTrigger>
+      <PopoverContent className="w-auto p-0" align="start">
+        <Calendar
+          mode="single"
+          selected={field.value}
+          onSelect={(date: Date | undefined) => field.onChange(date)}
+          disabled={(date) =>
+            date > new Date() || date < new Date("1900-01-01")
+          }
+          initialFocus
+        />
+      </PopoverContent>
+    </Popover>
+  );
+};
+
+export default PopoverComponent;
