+++ src/pages/home/index.tsx
@@ -6,11 +6,6 @@ import { TicketPromosBanner } from "./components/containers/TicketPromosBanner"
 import { Services } from "./components/containers/OurServices"
 import { NewsSection } from "./components/containers/NewsSection"
 import { CtaSection } from "./components/containers/CtaSection"
-// import NewsLetter from "../organs/NewsLetter"
-// import Partners from "../organs/Partners"
-// import Services from "../organs/Services"
-// import Testimonials from "../organs/Testimonials"
-// import TopDestination from "../organs/TopDestination"
 
 const Home = () => {
     return (
@@ -22,12 +17,6 @@ const Home = () => {
             <Services />
             <NewsSection />
             <CtaSection />
-            {/* <Services /> */}
-            {/* <TopDestination /> */}
-            {/* <BookingSteps /> */}
-            {/* <Testimonials /> */}
-            {/* <Partners /> */}
-            {/* <NewsLetter /> */}
         </>
     )
 }
